+++ R/printers.R
@@ -20,6 +20,7 @@ htmltools_value <- function(x, ft.align = NULL, ft.shadow = NULL,
                             extra_dependencies = NULL) {
   x <- flextable_global$defaults$post_process_all(x)
   x <- flextable_global$defaults$post_process_html(x)
+  x <- fix_border_issues(x)
 
   if (!is.null(ft.align)) {
     x$properties$align <- ft.align
@@ -143,6 +144,7 @@ to_html.flextable <- function(x, type = c("table", "img"), ...) {
   if ("table" %in% type_output) {
     x <- flextable_global$defaults$post_process_all(x)
     x <- flextable_global$defaults$post_process_html(x)
+    x <- fix_border_issues(x)
     manual_css <- readLines(system.file(package = "flextable", "web_1.1.3", "tabwid.css"), encoding = "UTF-8")
     gen_raw_html(x, class = "tabwid", caption = "", manual_css = paste0(manual_css, collapse = "\n"))
   } else {
@@ -181,6 +183,7 @@ to_html.flextable <- function(x, type = c("table", "img"), ...) {
 to_wml.flextable <- function(x, ...) {
   x <- flextable_global$defaults$post_process_all(x)
   x <- flextable_global$defaults$post_process_docx(x)
+  x <- fix_border_issues(x)
   x <- knitr_update_properties(x)
   gen_raw_wml(x)
 }
@@ -198,6 +201,7 @@ to_wml.flextable <- function(x, ...) {
 knit_to_html <- function(x, bookdown = FALSE, quarto = FALSE) {
   x <- flextable_global$defaults$post_process_all(x)
   x <- flextable_global$defaults$post_process_html(x)
+  x <- fix_border_issues(x)
 
   tab_props <- opts_current_table()
   topcaption <- tab_props$topcaption
@@ -231,6 +235,7 @@ knit_to_html <- function(x, bookdown = FALSE, quarto = FALSE) {
 knit_to_wml <- function(x, bookdown = FALSE, quarto = FALSE) {
   x <- flextable_global$defaults$post_process_all(x)
   x <- flextable_global$defaults$post_process_docx(x)
+  x <- fix_border_issues(x)
 
   is_rdocx_document <- opts_current$get("is_rdocx_document")
   if (is.null(is_rdocx_document)) is_rdocx_document <- FALSE
@@ -313,6 +318,7 @@ knit_to_latex <- function(x, bookdown, quarto = FALSE) {
 
   x <- flextable_global$defaults$post_process_all(x)
   x <- flextable_global$defaults$post_process_pdf(x)
+  x <- fix_border_issues(x)
 
   if ("none" %in% ft.latex.float) {
     lat_container <- latex_container_none()
@@ -387,6 +393,7 @@ knit_to_pml <- function(x) {
   }
   x <- flextable_global$defaults$post_process_all(x)
   x <- flextable_global$defaults$post_process_pptx(x)
+  x <- fix_border_issues(x)
 
   uid <- as.integer(runif(n = 1) * 10^9)
 
@@ -739,6 +746,7 @@ save_as_html <- function(..., values = NULL, path,
   values <- Filter(function(x) inherits(x, "flextable"), values)
   values <- lapply(values, flextable_global$defaults$post_process_all)
   values <- lapply(values, flextable_global$defaults$post_process_html)
+  values <- lapply(values, fix_border_issues)
   values <- lapply(values, htmltools_value)
   titles <- names(values)
   show_names <- !is.null(titles)