+++ NEWS.md
@@ -18,6 +18,7 @@ rmarkdown (issue #632)
 - `proc_freq` can now display only the table percentages without the count 
 using `include.table_count = FALSE`.
 - bring back support for 'pagedown' with `pagedown >= 0.20.2`
+- apply defined text-format to empty cells  
 
 # flextable 0.9.6