+++ NEWS.md
@@ -1,5 +1,9 @@
 # flextable 0.9.7
 
+## new features
+
+- Added support for labelled datasets.
+
 ## Changes
 
 - The `fix_border_issues` function is now useless for users, as it is now 
@@ -18,7 +22,7 @@ rmarkdown (issue #632)
 - `proc_freq` can now display only the table percentages without the count 
 using `include.table_count = FALSE`.
 - bring back support for 'pagedown' with `pagedown >= 0.20.2`
-- apply defined text-format to empty cells  
+- flextable now applies defined text-format to empty cells  
 
 # flextable 0.9.6